Moved Parser and Tokenizer to the Model
