+++ src/app/Model/Emulator/control-store.service.ts
@@ -1,4 +1,4 @@
-import { Token } from '../../Controller/micro-tokenizer.service';
+import { Token } from '../micro-tokenizer.service';
 import { Injectable } from '@angular/core';
 import { ParserService, Instruction, Line } from 'src/app/Model/Emulator/parser.service';
 import { MicroProviderService } from '../micro-provider.service';
