+++ src/app/Model/micro-tokenizer.service.ts
